requirements.txt
transformers==4.40.0
numpy==1.26.4
wandb==0.17.2
abctoolkit==0.0.6
samplings==0.1.7
pyparsing==3.2.1
gradio==5.17.1
tqdm==4.67.1
python-dotenv==1.0.1
